telegram-assistant/requirements.txt
python-telegram-bot==21.9
anthropic==0.49.0
python-dotenv==1.0.1
requests==2.32.3
